A lot of styling changes + unauthorized page redirect

+++ front-end/services/SeriesService.ts
@@ -1,19 +1,33 @@
 import { Series } from "@types";
 
 const getAllSeries = () => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+
     return fetch(process.env.NEXT_PUBLIC_API_URL + "/media/series", {
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         }
 })
 };
 
 const createSeries = async (newSeries : Series) => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+
     const response = await fetch(process.env.NEXT_PUBLIC_API_URL + "/media", {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         },
         body: JSON.stringify(newSeries)
 });
@@ -24,10 +38,17 @@ return response.json();
 };
 
 const deleteSeries = async (id: number) => {
+    const token = JSON.parse(sessionStorage.getItem("loggedInUser") as string).token;
+
+    if (!token) {
+        throw new Error("No token found");
+    }
+    
     const response = await fetch(process.env.NEXT_PUBLIC_API_URL + `/media/${id}`, {
         method: "DELETE",
         headers: {
             "Content-Type": "application/json",
+            Authorization: `Bearer ${token}`,
         }
 });
 if (!response.ok) {
